Move Options to options package
