+++ db.go
@@ -71,7 +71,7 @@ type DB struct {
 	elog      trace.EventLog
 	mt        *skl.Skiplist   // Our latest (actively written) in-memory table
 	imm       []*skl.Skiplist // Add here only AFTER pushing to flushChan.
-	opt       Options
+	opt       options.Options
 	manifest  *manifestFile
 	lc        *levelsController
 	vlog      valueLog
@@ -90,6 +90,16 @@ type DB struct {
 	orc *oracle
 
 	pub *publisher
+
+	// // Transaction start and commit timestamps are managed by end-user.
+	// // This is only useful for databases built on top of Badger (like Dgraph).
+	// // Not recommended for most users.
+	// managedTxns bool
+
+	// Flags for testing purposes
+	// ------------------------------
+	maxBatchCount int64 // max entries in batch
+	maxBatchSize  int64 // max batch size in bytes
 }
 
 const (
@@ -185,21 +195,21 @@ func (db *DB) replayFunction() func(Entry, valuePointer) error {
 }
 
 // Open returns a new DB object.
-func Open(opt Options) (db *DB, err error) {
-	opt.maxBatchSize = (15 * opt.MaxTableSize) / 100
-	opt.maxBatchCount = opt.maxBatchSize / int64(skl.MaxNodeSize)
+func Open(opt options.Options) (db *DB, err error) {
+	maxBatchSize := (15 * opt.MaxTableSize) / 100
+	maxBatchCount := maxBatchSize / int64(skl.MaxNodeSize)
 
 	// We are limiting opt.ValueThreshold to maxValueThreshold for now.
-	if opt.ValueThreshold > maxValueThreshold {
+	if opt.ValueThreshold > options.MaxValueThreshold {
 		return nil, errors.Errorf("Invalid ValueThreshold, must be less or equal to %d",
-			maxValueThreshold)
+			options.MaxValueThreshold)
 	}
 
 	// If ValueThreshold is greater than opt.maxBatchSize, we won't be able to push any data using
 	// the transaction APIs. Transaction batches entries into batches of size opt.maxBatchSize.
-	if int64(opt.ValueThreshold) > opt.maxBatchSize {
+	if int64(opt.ValueThreshold) > maxBatchSize {
 		return nil, errors.Errorf("Valuethreshold greater than max batch size of %d. Either "+
-			"reduce opt.ValueThreshold or increase opt.MaxTableSize.", opt.maxBatchSize)
+			"reduce opt.ValueThreshold or increase opt.MaxTableSize.", maxBatchSize)
 	}
 
 	if opt.ReadOnly {
@@ -282,13 +292,15 @@ func Open(opt Options) (db *DB, err error) {
 		valueDirGuard: valueDirLockGuard,
 		orc:           newOracle(opt),
 		pub:           newPublisher(),
+		maxBatchSize:  maxBatchSize,
+		maxBatchCount: maxBatchCount,
 	}
 
 	// Calculate initial size.
 	db.calculateSize()
 	db.closers.updateSize = y.NewCloser(1)
 	go db.updateSize(db.closers.updateSize)
-	db.mt = skl.NewSkiplist(arenaSize(opt))
+	db.mt = skl.NewSkiplist(db.arenaSize())
 
 	// newLevelsController potentially loads files in directory.
 	if db.lc, err = newLevelsController(db, &manifest); err != nil {
@@ -682,7 +694,7 @@ func (db *DB) sendToWriteCh(entries []*Entry) (*request, error) {
 		size += int64(e.estimateSize(db.opt.ValueThreshold))
 		count++
 	}
-	if count >= db.opt.maxBatchCount || size >= db.opt.maxBatchSize {
+	if count >= db.maxBatchCount || size >= db.maxBatchSize {
 		return nil, ErrTxnTooBig
 	}
 
@@ -835,7 +847,7 @@ func (db *DB) ensureRoomForWrite() error {
 			db.mt.MemSize(), len(db.flushChan))
 		// We manage to push this task. Let's modify imm.
 		db.imm = append(db.imm, db.mt)
-		db.mt = skl.NewSkiplist(arenaSize(db.opt))
+		db.mt = skl.NewSkiplist(db.arenaSize())
 		// New memtable is empty. We certainly have room.
 		return nil
 	default:
@@ -844,8 +856,8 @@ func (db *DB) ensureRoomForWrite() error {
 	}
 }
 
-func arenaSize(opt Options) int64 {
-	return opt.MaxTableSize + opt.maxBatchSize + opt.maxBatchCount*int64(skl.MaxNodeSize)
+func (db *DB) arenaSize() int64 {
+	return db.opt.MaxTableSize + db.maxBatchSize + db.maxBatchCount*int64(skl.MaxNodeSize)
 }
 
 // WriteLevel0Table flushes memtable.
@@ -1164,7 +1176,7 @@ func (seq *Sequence) updateLease() error {
 //
 // GetSequence is not supported on ManagedDB. Calling this would result in a panic.
 func (db *DB) GetSequence(key []byte, bandwidth uint64) (*Sequence, error) {
-	if db.opt.managedTxns {
+	if db.opt.ManagedTxns {
 		panic("Cannot use GetSequence with managedDB=true.")
 	}
 
@@ -1209,12 +1221,12 @@ func (db *DB) KeySplits(prefix []byte) []string {
 
 // MaxBatchCount returns max possible entries in batch
 func (db *DB) MaxBatchCount() int64 {
-	return db.opt.maxBatchCount
+	return db.maxBatchCount
 }
 
 // MaxBatchSize returns max possible batch size
 func (db *DB) MaxBatchSize() int64 {
-	return db.opt.maxBatchSize
+	return db.maxBatchSize
 }
 
 func (db *DB) stopMemoryFlush() {
@@ -1407,7 +1419,7 @@ func (db *DB) dropAll() (func(), error) {
 		mt.DecrRef()
 	}
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt)) // Set it up for future writes.
+	db.mt = skl.NewSkiplist(db.arenaSize()) // Set it up for future writes.
 
 	num, err := db.lc.dropTree()
 	if err != nil {
@@ -1465,7 +1477,7 @@ func (db *DB) DropPrefix(prefix []byte) error {
 	db.stopCompactions()
 	defer db.startCompactions()
 	db.imm = db.imm[:0]
-	db.mt = skl.NewSkiplist(arenaSize(db.opt))
+	db.mt = skl.NewSkiplist(db.arenaSize())
 
 	// Drop prefixes from the levels.
 	if err := db.lc.dropPrefix(prefix); err != nil {
