+++ logger.go
@@ -16,70 +16,70 @@
 
 package badger
 
-import (
-	"log"
-	"os"
-)
+// import (
+// 	"log"
+// 	"os"
+// )
 
-// Logger is implemented by any logging system that is used for standard logs.
-type Logger interface {
-	Errorf(string, ...interface{})
-	Warningf(string, ...interface{})
-	Infof(string, ...interface{})
-	Debugf(string, ...interface{})
-}
+// // Logger is implemented by any logging system that is used for standard logs.
+// type Logger interface {
+// 	Errorf(string, ...interface{})
+// 	Warningf(string, ...interface{})
+// 	Infof(string, ...interface{})
+// 	Debugf(string, ...interface{})
+// }
 
-// Errorf logs an ERROR log message to the logger specified in opts or to the
-// global logger if no logger is specified in opts.
-func (opt *Options) Errorf(format string, v ...interface{}) {
-	if opt.Logger == nil {
-		return
-	}
-	opt.Logger.Errorf(format, v...)
-}
+// // Errorf logs an ERROR log message to the logger specified in opts or to the
+// // global logger if no logger is specified in opts.
+// func (opt *Options) Errorf(format string, v ...interface{}) {
+// 	if opt.Logger == nil {
+// 		return
+// 	}
+// 	opt.Logger.Errorf(format, v...)
+// }
 
-// Infof logs an INFO message to the logger specified in opts.
-func (opt *Options) Infof(format string, v ...interface{}) {
-	if opt.Logger == nil {
-		return
-	}
-	opt.Logger.Infof(format, v...)
-}
+// // Infof logs an INFO message to the logger specified in opts.
+// func (opt *Options) Infof(format string, v ...interface{}) {
+// 	if opt.Logger == nil {
+// 		return
+// 	}
+// 	opt.Logger.Infof(format, v...)
+// }
 
-// Warningf logs a WARNING message to the logger specified in opts.
-func (opt *Options) Warningf(format string, v ...interface{}) {
-	if opt.Logger == nil {
-		return
-	}
-	opt.Logger.Warningf(format, v...)
-}
+// // Warningf logs a WARNING message to the logger specified in opts.
+// func (opt *Options) Warningf(format string, v ...interface{}) {
+// 	if opt.Logger == nil {
+// 		return
+// 	}
+// 	opt.Logger.Warningf(format, v...)
+// }
 
-// Debugf logs a DEBUG message to the logger specified in opts.
-func (opt *Options) Debugf(format string, v ...interface{}) {
-	if opt.Logger == nil {
-		return
-	}
-	opt.Logger.Debugf(format, v...)
-}
+// // Debugf logs a DEBUG message to the logger specified in opts.
+// func (opt *Options) Debugf(format string, v ...interface{}) {
+// 	if opt.Logger == nil {
+// 		return
+// 	}
+// 	opt.Logger.Debugf(format, v...)
+// }
 
-type defaultLog struct {
-	*log.Logger
-}
+// type defaultLog struct {
+// 	*log.Logger
+// }
 
-var defaultLogger = &defaultLog{Logger: log.New(os.Stderr, "badger ", log.LstdFlags)}
+// var defaultLogger = &defaultLog{Logger: log.New(os.Stderr, "badger ", log.LstdFlags)}
 
-func (l *defaultLog) Errorf(f string, v ...interface{}) {
-	l.Printf("ERROR: "+f, v...)
-}
+// func (l *defaultLog) Errorf(f string, v ...interface{}) {
+// 	l.Printf("ERROR: "+f, v...)
+// }
 
-func (l *defaultLog) Warningf(f string, v ...interface{}) {
-	l.Printf("WARNING: "+f, v...)
-}
+// func (l *defaultLog) Warningf(f string, v ...interface{}) {
+// 	l.Printf("WARNING: "+f, v...)
+// }
 
-func (l *defaultLog) Infof(f string, v ...interface{}) {
-	l.Printf("INFO: "+f, v...)
-}
+// func (l *defaultLog) Infof(f string, v ...interface{}) {
+// 	l.Printf("INFO: "+f, v...)
+// }
 
-func (l *defaultLog) Debugf(f string, v ...interface{}) {
-	l.Printf("DEBUG: "+f, v...)
-}
+// func (l *defaultLog) Debugf(f string, v ...interface{}) {
+// 	l.Printf("DEBUG: "+f, v...)
+// }
